Demo.Mmose.Core.Test/ServerEngine/World/ZoneCluster/ZoneClusterTest.cs
using Demo.Mmose.Core.World.ZoneClusterWorld;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Demo.Mmose.Core.World;
using Demo.Mmose.Core.Network;
using Demo.Mmose.Core.World.DomainWorld;

namespace Demo.Mmose.Core.Test
{
    /// <summary>
    ///这是 ZoneClusterTest 的测试类，旨在
    ///包含所有 ZoneClusterTest 单元测试
    ///</summary>
    [TestClass()]
    public class ZoneClusterTest
    {
        private TestContext testContextInstance;
        /// <summary>
        ///获取或设置测试上下文，上下文提供
        ///有关当前测试运行及其功能的信息。
        ///</summary>
        public TestContext TestContext
        {
            get { return testContextInstance; }
            set { testContextInstance = value; }
        }

        #region 附加测试属性
        // 
        //编写测试时，还可使用以下属性:
        //
        //使用 ClassInitialize 在运行类中的第一个测试前先运行代码
        //[ClassInitialize()]
        //public static void MyClassInitialize(TestContext testContext)
        //{
        //}
        //
        //使用 ClassCleanup 在运行完类中的所有测试后再运行代码
        //[ClassCleanup()]
        //public static void MyClassCleanup()
        //{
        //}
        //
        //使用 TestInitialize 在运行每个测试前先运行代码
        //[TestInitialize()]
        //public void MyTestInitialize()
        //{
        //}
        //
        //使用 TestCleanup 在运行完每个测试后运行代码
        //[TestCleanup()]
        //public void MyTestCleanup()
        //{
        //}
        //
        #endregion

        /// <summary>
        ///ZoneHandler 的测试
        ///</summary>
        public void ZoneHandlerTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            ZoneHandler actual;
            actual = target.ZoneHandler;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ZoneHandlerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneHandlerTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ZoneClusterPacketHandlers 的测试
        ///</summary>
        public void ZoneClusterPacketHandlersTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            ZoneClusterPacketHandlers actual;
            actual = target.ZoneClusterPacketHandlers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ZoneClusterPacketHandlersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneClusterPacketHandlersTestHelper<Wor" +
                    "ldT>()。" );
        }

        /// <summary>
        ///World 的测试
        ///</summary>
        public void WorldTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            WorldT actual;
            actual = target.World;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void WorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 WorldTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ProcessorBuffers 的测试
        ///</summary>
        public void ProcessorBuffersTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            BufferPool actual;
            actual = target.ProcessorBuffers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ProcessorBuffersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ProcessorBuffersTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///IsLoginDomain 的测试
        ///</summary>
        public void IsLoginDomainTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            bool actual;
            actual = target.IsLoginDomain;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void IsLoginDomainTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 IsLoginDomainTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///DomainHandler 的测试
        ///</summary>
        public void DomainHandlerTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            ZoneClusterHandler actual;
            actual = target.DomainHandler;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void DomainHandlerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 DomainHandlerTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ConfigZoneCluster 的测试
        ///</summary>
        public void ConfigZoneClusterTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            ConfigZoneCluster actual;
            actual = target.ConfigZoneCluster;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ConfigZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ConfigZoneClusterTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ZoneClusterInitOnceServer 的测试
        ///</summary>
        public void ZoneClusterInitOnceServerTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ZoneClusterInitOnceServer( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ZoneClusterInitOnceServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneClusterInitOnceServerTestHelper<Wor" +
                    "ldT>()。" );
        }

        /// <summary>
        ///ZoneClusterExitWorld 的测试
        ///</summary>
        public void ZoneClusterExitWorldTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ZoneClusterExitWorld( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ZoneClusterExitWorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneClusterExitWorldTestHelper<WorldT>(" +
                    ")。" );
        }

        /// <summary>
        ///unInitZoneCluster 的测试
        ///</summary>
        public void unInitZoneClusterTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            target.unInitZoneCluster();
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void unInitZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 unInitZoneClusterTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///SentToZone 的测试
        ///</summary>
        public void SentToZoneTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            NetState toNetState = null; // TODO: 初始化为适当的值
            string strMessage = string.Empty; // TODO: 初始化为适当的值
            int expected = 0; // TODO: 初始化为适当的值
            int actual;
            actual = target.SentToZone( toNetState, strMessage );
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void SentToZoneTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 SentToZoneTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///OnLoginDomain 的测试
        ///</summary>
        public void OnLoginDomainTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateConnectEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.OnLoginDomain( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void OnLoginDomainTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 OnLoginDomainTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ListenerProcessReceive 的测试
        ///</summary>
        public void ListenerProcessReceiveTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ListenerProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ListenerProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerProcessReceiveTestHelper<WorldT" +
                    ">()。" );
        }

        /// <summary>
        ///ListenerInitNetState 的测试
        ///</summary>
        public void ListenerInitNetStateTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ListenerInitNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ListenerInitNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerInitNetStateTestHelper<WorldT>(" +
                    ")。" );
        }

        /// <summary>
        ///InitZoneCluster 的测试
        ///</summary>
        public void InitZoneClusterTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            ConfigZoneCluster configZoneCluster = null; // TODO: 初始化为适当的值
            target.InitZoneCluster( configZoneCluster );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void InitZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 InitZoneClusterTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///GetZoneId 的测试
        ///</summary>
        public void GetZoneIdTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            NetState netState = null; // TODO: 初始化为适当的值
            ZoneCluster<WorldT>.ZoneID expected = null; // TODO: 初始化为适当的值
            ZoneCluster<WorldT>.ZoneID actual;
            actual = target.GetZoneId( netState );
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void GetZoneIdTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 GetZoneIdTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///GetPacketLength 的测试
        ///</summary>
        public void GetPacketLengthTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketLengthInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketLength( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketLengthTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketLengthTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///GetPacketID 的测试
        ///</summary>
        public void GetPacketIDTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketIdInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketID( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketIDTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketIDTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///CreateChannel 的测试
        ///</summary>
        public void CreateChannelTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            ZoneCluster<WorldT>.Channel expected = null; // TODO: 初始化为适当的值
            ZoneCluster<WorldT>.Channel actual;
            actual = target.CreateChannel();
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void CreateChannelTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 CreateChannelTestHelper<WorldT>()。" );
        }

        /// <summary>
        ///ConnectToDomainServer 的测试
        ///</summary>
        public void ConnectToDomainServerTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster_Accessor<WorldT> target = new ZoneCluster_Accessor<WorldT>(); // TODO: 初始化为适当的值
            bool expected = false; // TODO: 初始化为适当的值
            bool actual;
            actual = target.ConnectToDomainServer();
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ConnectToDomainServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnectToDomainServerTestHelper<WorldT>" +
                    "()。" );
        }

        /// <summary>
        ///ConnecterProcessReceive 的测试
        ///</summary>
        public void ConnecterProcessReceiveTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ConnecterProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ConnecterProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterProcessReceiveTestHelper<World" +
                    "T>()。" );
        }

        /// <summary>
        ///ConnecterInitNetState 的测试
        ///</summary>
        public void ConnecterInitNetStateTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ConnecterInitNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ConnecterInitNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterInitNetStateTestHelper<WorldT>" +
                    "()。" );
        }

        /// <summary>
        ///ZoneCluster`1 构造函数 的测试
        ///</summary>
        public void ZoneClusterConstructorTestHelper<WorldT>()
            where WorldT : BaseWorld, new()
        {
            ZoneCluster<WorldT> target = new ZoneCluster<WorldT>();
            Assert.Inconclusive( "TODO: 实现用来验证目标的代码" );
        }

        [TestMethod()]
        public void ZoneClusterConstructorTest()
        {
            Assert.Inconclusive( "没有找到能够满足 WorldT 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneClusterConstructorTestHelper<WorldT" +
                    ">()。" );
        }
    }
}
